修改播放器 前进的生命周期

+++ pages/playaudio/playaudio.vue
@@ -212,10 +212,11 @@ export default {
 			this.isFinish = true;
 		}
 	},
-	
+	onHide() {
+		 if (this.audioContext) this.audioContext.destroy();
+	},
 	onUnload() {
-		// console.log("audio",this.audioContext);
-		this.audioContext.destroy();
+		if (this.audioContext) this.audioContext.destroy();
 		uni.setStorageSync('flag',true)
 	},
 	components: {
